+++ mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/text.kt
@@ -4,6 +4,8 @@ package com.jakewharton.mosaic
 
 import androidx.compose.runtime.Stable
 
+internal const val SpaceCharCodePoint = ' '.code
+
 /**
  * Unicode code point cannot contain a negative value.
  */