+++ mosaic-runtime/src/commonTest/kotlin/com/jakewharton/mosaic/CounterTest.kt
@@ -35,7 +35,7 @@ class CounterTest {
 			terminalState.update { copy(size = IntSize(width = 30, height = 1)) }
 			setCounterInTerminalCenter()
 			for (count in 0..9) {
-				assertThat(awaitSnapshot()).isEqualTo("        The count is: $count       ")
+				assertThat(awaitSnapshot()).isEqualTo("        The count is: $count")
 			}
 
 			terminalState.update { copy(size = IntSize(width = 20, height = 1)) }
@@ -45,7 +45,7 @@ class CounterTest {
 			delay(250L)
 
 			for (count in 10..20) {
-				assertThat(awaitSnapshot()).isEqualTo("  The count is: $count  ")
+				assertThat(awaitSnapshot()).isEqualTo("  The count is: $count")
 			}
 		}
 	}
@@ -60,7 +60,7 @@ class CounterTest {
 			for (count in 0..20) {
 				assertThat(awaitSnapshot()).isEqualTo(
 					"""
-					|The count is: $count      $s
+					|The count is: $count
 					|The second count is: $count
 					""".trimMargin(),
 				)