+++ mosaic-runtime/src/commonTest/kotlin/com/jakewharton/mosaic/MosaicTest.kt
@@ -46,8 +46,8 @@ class MosaicTest {
 				}
 			}
 
-			assertThat(awaitSnapshot()).isEqualTo("$TestChar         ")
-			assertThat(awaitSnapshot()).isEqualTo("     $TestChar    ")
+			assertThat(awaitSnapshot()).isEqualTo("$TestChar")
+			assertThat(awaitSnapshot()).isEqualTo("     $TestChar")
 		}
 	}
 
@@ -66,8 +66,8 @@ class MosaicTest {
 				}
 			}
 
-			assertThat(awaitSnapshot()).isEqualTo("$TestChar         ")
-			assertThat(awaitSnapshot()).isEqualTo("     $TestChar    ")
+			assertThat(awaitSnapshot()).isEqualTo("$TestChar")
+			assertThat(awaitSnapshot()).isEqualTo("     $TestChar")
 		}
 	}
 
@@ -96,8 +96,8 @@ class MosaicTest {
 				}
 			}
 
-			assertThat(awaitSnapshot()).isEqualTo("$TestChar         ")
-			assertThat(awaitSnapshot()).isEqualTo("${TestChar + 1}         ")
+			assertThat(awaitSnapshot()).isEqualTo("$TestChar")
+			assertThat(awaitSnapshot()).isEqualTo("${TestChar + 1}")
 		}
 	}