+++ mosaic-runtime/src/commonTest/kotlin/com/jakewharton/mosaic/layout/OffsetTest.kt
@@ -6,7 +6,6 @@ import com.jakewharton.mosaic.TestChar
 import com.jakewharton.mosaic.TestFiller
 import com.jakewharton.mosaic.assertFailure
 import com.jakewharton.mosaic.modifier.Modifier
-import com.jakewharton.mosaic.s
 import com.jakewharton.mosaic.testing.runMosaicTest
 import com.jakewharton.mosaic.ui.Box
 import com.jakewharton.mosaic.ui.unit.IntOffset
@@ -23,12 +22,12 @@ class OffsetTest {
 			}
 			assertThat(awaitSnapshot()).isEqualTo(
 				"""
-				|   $TestChar $s
-				|     $s
-				|     $s
-				|     $s
-				|     $s
-				|     $s
+				|   $TestChar
+				|
+				|
+				|
+				|
+				|
 				""".trimMargin(),
 			)
 		}
@@ -71,12 +70,12 @@ class OffsetTest {
 			}
 			assertThat(awaitSnapshot()).isEqualTo(
 				"""
-				|     $s
-				|     $s
-				|     $s
-				|     $s
-				|$TestChar    $s
-				|     $s
+				|
+				|
+				|
+				|
+				|$TestChar
+				|
 				""".trimMargin(),
 			)
 		}
@@ -119,12 +118,12 @@ class OffsetTest {
 			}
 			assertThat(awaitSnapshot()).isEqualTo(
 				"""
-				|     $s
-				|     $s
-				|     $s
-				|     $s
-				|   $TestChar $s
-				|     $s
+				|
+				|
+				|
+				|
+				|   $TestChar
+				|
 				""".trimMargin(),
 			)
 		}
@@ -172,12 +171,12 @@ class OffsetTest {
 			}
 			assertThat(awaitSnapshot()).isEqualTo(
 				"""
-				|   $TestChar $s
-				|     $s
-				|     $s
-				|     $s
-				|     $s
-				|     $s
+				|   $TestChar
+				|
+				|
+				|
+				|
+				|
 				""".trimMargin(),
 			)
 		}
@@ -220,12 +219,12 @@ class OffsetTest {
 			}
 			assertThat(awaitSnapshot()).isEqualTo(
 				"""
-				|     $s
-				|     $s
-				|     $s
-				|     $s
-				|$TestChar    $s
-				|     $s
+				|
+				|
+				|
+				|
+				|$TestChar
+				|
 				""".trimMargin(),
 			)
 		}
@@ -268,12 +267,12 @@ class OffsetTest {
 			}
 			assertThat(awaitSnapshot()).isEqualTo(
 				"""
-				|     $s
-				|     $s
-				|     $s
-				|     $s
-				|   $TestChar $s
-				|     $s
+				|
+				|
+				|
+				|
+				|   $TestChar
+				|
 				""".trimMargin(),
 			)
 		}