+++ mosaic-runtime/src/commonTest/kotlin/com/jakewharton/mosaic/ui/FillerTest.kt
@@ -12,7 +12,6 @@ import com.jakewharton.mosaic.layout.padding
 import com.jakewharton.mosaic.layout.size
 import com.jakewharton.mosaic.layout.width
 import com.jakewharton.mosaic.modifier.Modifier
-import com.jakewharton.mosaic.s
 import com.jakewharton.mosaic.size
 import com.jakewharton.mosaic.testing.runMosaicTest
 import com.jakewharton.mosaic.ui.unit.Constraints
@@ -52,12 +51,12 @@ class FillerTest {
 			}
 			assertThat(awaitSnapshot()).isEqualTo(
 				"""
-				|   $s
-				| $TestChar$TestChar$s
-				| $TestChar$TestChar$s
-				| $TestChar$TestChar$s
-				| $TestChar$TestChar$s
-				|   $s
+				|
+				| $TestChar$TestChar
+				| $TestChar$TestChar
+				| $TestChar$TestChar
+				| $TestChar$TestChar
+				|
 				""".trimMargin(),
 			)
 		}